+++ src/twin/mod.rs
@@ -16,6 +16,7 @@ use crate::twin::{
     network_status::NetworkStatus, reboot::Reboot, ssh_tunnel::SshTunnel,
     wifi_commissioning::WifiCommissioning,
 };
+use crate::update_validation::UpdateValidation;
 use anyhow::{anyhow, bail, Result};
 use async_trait::async_trait;
 use azure_iot_sdk::client::*;
@@ -90,10 +91,11 @@ pub struct Twin {
     rx_reported_properties: mpsc::Receiver<serde_json::Value>,
     rx_outgoing_message: mpsc::Receiver<IotMessage>,
     features: HashMap<TypeId, Box<dyn Feature>>,
+    update_validation: update_validation::UpdateValidation,
 }
 
 impl Twin {
-    pub fn new(client: Box<dyn IotHub>) -> Self {
+    pub fn new(client: Box<dyn IotHub>, update_validation: UpdateValidation) -> Self {
         let (tx_reported_properties, rx_reported_properties) = mpsc::channel(100);
         let (tx_outgoing_message, rx_outgoing_message) = mpsc::channel(100);
 
@@ -136,6 +138,7 @@ impl Twin {
             rx_outgoing_message,
             authenticated_once: false,
             features,
+            update_validation,
         }
     }
 
@@ -222,8 +225,6 @@ impl Twin {
 
         match auth_status {
             AuthenticationStatus::Authenticated => {
-                let mut update_validated: Result<()> = Ok(());
-
                 if !self.authenticated_once {
                     /*
                      * the update validation test "wait_for_system_running" enforces that
@@ -231,14 +232,13 @@ impl Twin {
                      */
 
                     systemd::sd_notify_ready();
-                    update_validated = update_validation::check().await;
+
+                    self.update_validation.set_authenticated().await?;
 
                     self.init().await?;
 
                     self.authenticated_once = true;
                 };
-
-                update_validated?;
             }
             AuthenticationStatus::Unauthenticated(reason) => {
                 anyhow::ensure!(
@@ -334,6 +334,10 @@ impl Twin {
             None
         };
 
+        // has to be called before iothub client authentication
+        let update_validation = UpdateValidation::new()?;
+
+        info!("waiting for authentication...");
         let client = match IotHubClient::client_type() {
             _ if connection_string.is_some() => IotHubClient::from_connection_string(
                 connection_string.unwrap(),
@@ -359,7 +363,7 @@ impl Twin {
             )?,
         };
 
-        let mut twin = Self::new(client);
+        let mut twin = Self::new(client, update_validation);
         let handle = signals.handle();
 
         loop {