+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] Q2 2024
+- refactoring of update validation handling:
+  - introduced post service script to handle omnect-device-service restarts during
+    update_validation
+  - global timeout for update validation
+- updated azure-iot-sdk to 0.11.11 (confirmation timeout increased + do log instead panic on timeout)
+
 ## [0.16.2] Q2 2024
 - fixed Cargo.lock